requirements.txt
appdirs==1.4.4
asttokens==3.0.0
asyncio==3.4.3
bandit==1.8.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
debugpy==1.8.11
dnspython==2.7.0
eventlet==0.38.2
executing==2.1.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
greenlet==3.1.1
h11==0.14.0
icecream==2.1.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pbr==6.1.0
Pygments==2.18.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
queuelib==1.7.0
requests==2.32.3
rich==13.9.4
simple-websocket==1.1.0
six==1.17.0
socketIO-client==0.7.2
stevedore==5.4.0
urllib3==2.2.3
uWSGI==2.0.28
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
